+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-04-23 21:43:40
-// Commit: fix(ui): fix memory leak by cancelling ui subscription
+// Updated at: 2026-04-23 22:22:13
+// Commit: chore(ui): update dependencies and cleanup unused imports in ui
 
 class uiComponentState {
-  final String commitRef = '618';
-  final String lastMessage = 'fix(ui): fix memory leak by cancelling ui subscription';
-  final DateTime timestamp = DateTime.parse('2026-04-23T21:43:40+05:00');
+  final String commitRef = '621';
+  final String lastMessage = 'chore(ui): update dependencies and cleanup unused imports in ui';
+  final DateTime timestamp = DateTime.parse('2026-04-23T22:22:13+05:00');
 }